Refactored mail and database connection and used Settings.LIVE in mailer
