+++ faq-site-server/src/components/pages/user/userDashboard.ts
@@ -1,7 +1,7 @@
 import {app} from "../../../";
 import {UserDashboardCallBack, UserDashboardRequest} from "../../../../../faq-site-shared/api-calls";
 import {Request, Response} from "express";
-import {DatabaseResultSet, query} from "../../../database-connection";
+import {DatabaseResultSet, query} from "../../../utilities/database-connection";
 import {checkTokenValidity} from "../../../auth/middleware";
 
 app.post(UserDashboardRequest.getURL, (req: Request, res: Response) => {
